+++ intelmq/lib/harmonization.py
@@ -135,6 +135,9 @@ class Boolean(GenericType):
 
 
 class ClassificationType(GenericType):
+    """
+    classification.type type. Allowed values are:
+     * """
 
     allowed_values = ['spam',
                       'malware',
@@ -158,8 +161,12 @@ class ClassificationType(GenericType):
                       'blacklist',
                       'other',
                       'unknown',
-                      'test'
+                      'test',
+                      'tor',
+                      'leak',
                       ]
+
+    __doc__ += '\n     * '.join(allowed_values)
 
     @staticmethod
     def is_valid(value, sanitize=False):