+++ CHANGELOG.md
@@ -19,6 +19,7 @@ CHANGELOG
 - Modify Bot default ruleset: changed conficker rule to catch more spellings
 - Shadowserver Parser: Add Accessible Cisco Smart Install
 - CleanMX Phising Parser: Handle new columns `first` and `last` (#1131).
+- n6 parser: Fix classification mappings. See NEWS file for changes values. (#738)
 
 ### Documentation